feat: Adding the ability to enable and disable base and alternate recipes.
Added simplebar to display a custom scrollbar inside the recipe tables. Also added a few components to display the recipe selection within a data table.  Also added tabs to the production overview for the recipe options to live. Finally added a store for the recipe data and support for local storage.